[LEC] move local_client.rs to local.rs
